sandbox/src/main/java/ru/training/geometry/figures/Triangle.java
package ru.training.geometry.figures;

public record Triangle (double a, double b, double c) {
    public static void printTriangleArea(Triangle t) {
        String text = String.format("Площадь треугольника со сторонами %f, %f и %f = %f", t.a, t.b, t.c, t.area());
        System.out.println(text);
    }

    public double area() {
        var p = (this.a + this.b + this.c) / 2; // вычисление полупериметра
        return Math.sqrt(p * (p - this.a) * (p - this.b) * (p - this.c));
    }

    public double perimeter() {
        return this.a + this.b + this.c;
    }
}
